requirements.txt
#python version 3.10.6
Django==4.1
djangorestframework